refactor: restructure section components (collection, style) and update import paths
